pyproject.toml
[project]
name = "reffy"
version = "0.1.0"
description = "Local-first references store with Linear sync (MVP scaffold)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "starlette>=0.37",
  "uvicorn>=0.29",
  "httpx>=0.27",
  "python-dotenv>=1.0",
  "watchdog>=4.0",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
